moved LazyOneWayGraph to utils and renamed without underscore, updatd docs
